continuing improve
fixed some bugs
continue to verify starting from first call of (cd) in function main()
