+++ src/rover.c
@@ -6,7 +6,7 @@ void init_marks(Marks *marks)
 	strcpy(marks->dirpath, "");
 	marks->bulk     = BULK_INIT;
 	marks->nentries = 0;
-	marks->entries  = calloc(marks->bulk, sizeof *marks->entries);
+	marks->entries  = (char **)calloc(marks->bulk, sizeof(*marks->entries));
 }
 
 /* Unmark all entries. */
@@ -17,13 +17,12 @@ void mark_none(Marks *marks)
 	strcpy(marks->dirpath, "");
 	for (i = 0; i < marks->bulk && marks->nentries; i++)
 		if (marks->entries[i]) {
-			free(marks->entries[i]);
-			marks->entries[i] = NULL;
+			FREE(marks->entries[i]);
 			marks->nentries--;
 		}
 	if (marks->bulk > BULK_THRESH) {
 		/* Reset bulk to free some memory. */
-		free(marks->entries);
+		FREE(marks->entries);
 		marks->bulk    = BULK_INIT;
 		marks->entries = calloc(marks->bulk, sizeof *marks->entries);
 	}
@@ -69,8 +68,7 @@ void del_mark(Marks *marks, char *entry)
 		for (i = 0; i < marks->bulk; i++)
 			if (marks->entries[i] && !strcmp(marks->entries[i], entry))
 				break;
-		free(marks->entries[i]);
-		marks->entries[i] = NULL;
+		FREE(marks->entries[i]);
 		marks->nentries--;
 	} else
 		mark_none(marks);
@@ -82,10 +80,10 @@ void free_marks(Marks *marks)
 
 	for (i = 0; i < marks->bulk && marks->nentries; i++)
 		if (marks->entries[i]) {
-			free(marks->entries[i]);
+			FREE(marks->entries[i]);
 			marks->nentries--;
 		}
-	free(marks->entries);
+	FREE(marks->entries);
 }
 
 void handle_usr1(int sig)
@@ -214,9 +212,9 @@ int open_with_env(char *program, char *path)
 
 	if (program) {
 #ifdef RV_SHELL
-		strncpy(buffer, program, PATH_MAX -1);
+		strncpy(buffer, program, PATH_MAX - 1);
 		strncat(buffer, " ", PATH_MAX - strlen(program) - 1);
-		shell_escaped_cat(buffer, path, PATH_MAX - strlen(program) -2);
+		shell_escaped_cat(buffer, path, PATH_MAX - strlen(program) - 2);
 		spawn((char *[]){ RV_SHELL, "-c", buffer, NULL });
 #else
 		spawn((char *[]){ program, path, NULL });
@@ -224,123 +222,6 @@ int open_with_env(char *program, char *path)
 		return 1;
 	}
 	return 0;
-}
-
-
-/* Update the listing view. */
-void update_view()
-{
-	int i, j, numsize, ishidden, marking;
-	char buffer_one[PATH_MAX], buffer_two[PATH_MAX];
-	wchar_t wbuffer[PATH_MAX];
-
-	mvhline(0, 0, ' ', COLS);
-	attr_on(A_BOLD, NULL);
-	color_set(RVC_TABNUM, NULL);
-	mvaddch(0, COLS - 2, rover.tab + '0');
-	attr_off(A_BOLD, NULL);
-	if (rover.marks.nentries) {
-		numsize = snprintf(buffer_one, PATH_MAX, "%d", rover.marks.nentries);
-		color_set(RVC_MARKS, NULL);
-		mvaddstr(0, COLS - 3 - numsize, buffer_one);
-	} else
-		numsize = -1;
-	color_set(RVC_CWD, NULL);
-	mbstowcs(wbuffer, CWD, PATH_MAX);
-	mvaddnwstr(0, 0, wbuffer, COLS - 4 - numsize);
-	wcolor_set(rover.window, RVC_BORDER, NULL);
-	wborder(rover.window, 0, 0, 0, 0, 0, 0, 0, 0);
-	ESEL = MAX(MIN(ESEL, rover.nfiles - 1), 0);
-	/* Selection might not be visible, due to cursor wrapping or window
-       shrinking. In that case, the scroll must be moved to make it visible. */
-	if (rover.nfiles > HEIGHT) {
-		SCROLL = MAX(MIN(SCROLL, ESEL), ESEL - HEIGHT + 1);
-		SCROLL = MIN(MAX(SCROLL, 0), rover.nfiles - HEIGHT);
-	} else
-		SCROLL = 0;
-	marking = !strcmp(CWD, rover.marks.dirpath);
-	for (i = 0, j = SCROLL; i < HEIGHT && j < rover.nfiles; i++, j++) {
-		ishidden = ENAME(j)[0] == '.';
-		if (j == ESEL)
-			wattr_on(rover.window, A_REVERSE, NULL);
-		if (ISLINK(j))
-			wcolor_set(rover.window, RVC_LINK, NULL);
-		else if (ishidden)
-			wcolor_set(rover.window, RVC_HIDDEN, NULL);
-		else if (S_ISREG(EMODE(j))) {
-			if (EMODE(j) & (S_IXUSR | S_IXGRP | S_IXOTH))
-				wcolor_set(rover.window, RVC_EXEC, NULL);
-			else
-				wcolor_set(rover.window, RVC_REG, NULL);
-		} else if (S_ISDIR(EMODE(j)))
-			wcolor_set(rover.window, RVC_DIR, NULL);
-		else if (S_ISCHR(EMODE(j)))
-			wcolor_set(rover.window, RVC_CHR, NULL);
-		else if (S_ISBLK(EMODE(j)))
-			wcolor_set(rover.window, RVC_BLK, NULL);
-		else if (S_ISFIFO(EMODE(j)))
-			wcolor_set(rover.window, RVC_FIFO, NULL);
-		else if (S_ISSOCK(EMODE(j)))
-			wcolor_set(rover.window, RVC_SOCK, NULL);
-		if (S_ISDIR(EMODE(j))) {
-			mbstowcs(wbuffer, ENAME(j), PATH_MAX);
-			if (ISLINK(j))
-				wcscat(wbuffer, L"/");
-		} else {
-			char *suffix, *suffixes = "BKMGTPEZY";
-			off_t human_size = ESIZE(j) * 10;
-			int length       = mbstowcs(wbuffer, ENAME(j), PATH_MAX);
-			int namecols     = wcswidth(wbuffer, length);
-			for (suffix = suffixes; human_size >= 10240; suffix++)
-				human_size = (human_size + 512) / 1024;
-			if (*suffix == 'B')
-				swprintf(wbuffer + length, PATH_MAX - length, L"%*d %c",
-				         (int)(COLS - namecols - 6),
-				         (int)human_size / 10, *suffix);
-			else
-				swprintf(wbuffer + length, PATH_MAX - length, L"%*d.%d %c",
-				         (int)(COLS - namecols - 8),
-				         (int)human_size / 10, (int)human_size % 10, *suffix);
-		}
-		mvwhline(rover.window, i + 1, 1, ' ', COLS - 2);
-		mvwaddnwstr(rover.window, i + 1, 2, wbuffer, COLS - 4);
-		if (marking && MARKED(j)) {
-			wcolor_set(rover.window, RVC_MARKS, NULL);
-			mvwaddch(rover.window, i + 1, 1, RVS_MARK);
-		} else
-			mvwaddch(rover.window, i + 1, 1, ' ');
-		if (j == ESEL)
-			wattr_off(rover.window, A_REVERSE, NULL);
-	}
-	for (; i < HEIGHT; i++)
-		mvwhline(rover.window, i + 1, 1, ' ', COLS - 2);
-	if (rover.nfiles > HEIGHT) {
-		int center, height;
-		center = (SCROLL + HEIGHT / 2) * HEIGHT / rover.nfiles;
-		height = (HEIGHT - 1) * HEIGHT / rover.nfiles;
-		if (!height)
-			height = 1;
-		wcolor_set(rover.window, RVC_SCROLLBAR, NULL);
-		mvwvline(rover.window, center - height / 2 + 1, COLS - 1, RVS_SCROLLBAR, height);
-	}
-	buffer_one[0] = FLAGS & SHOW_FILES ? 'F' : ' ';
-	buffer_one[1] = FLAGS & SHOW_DIRS ? 'D' : ' ';
-	buffer_one[2] = FLAGS & SHOW_HIDDEN ? 'H' : ' ';
-	if (!rover.nfiles)
-		strcpy(buffer_two, "0/0");
-	else
-		snprintf(buffer_two, PATH_MAX, "%d/%d", ESEL + 1, rover.nfiles);
-	snprintf(buffer_one + 3, PATH_MAX, "%12s", buffer_two);
-	color_set(RVC_STATUS, NULL);
-	mvaddstr(LINES - 1, STATUSPOS, buffer_one);
-	wrefresh(rover.window);
-}
-
-
-/* Clear message area, leaving only status info. */
-void clear_message()
-{
-	mvhline(LINES - 1, 0, ' ', STATUSPOS);
 }
 
 /* Comparison used to sort listing entries. */
@@ -366,15 +247,19 @@ int ls(Row **rowsp, uint8_t flags)
 
 	if (!(dp = opendir(".")))
 		return -1;
+
 	n = -2; /* We don't want the entries "." and "..". */
 	while (readdir(dp))
 		n++;
+
 	if (n == 0) {
 		closedir(dp);
+
 		return 0;
 	}
+
 	rewinddir(dp);
-	rows = malloc(n * sizeof *rows);
+	rows = malloc(n * sizeof(*rows));
 	i    = 0;
 	while ((ep = readdir(dp))) {
 		if (!strcmp(ep->d_name, ".") || !strcmp(ep->d_name, ".."))
@@ -386,15 +271,15 @@ int ls(Row **rowsp, uint8_t flags)
 		stat(ep->d_name, &statbuf);
 		if (S_ISDIR(statbuf.st_mode)) {
 			if (flags & SHOW_DIRS) {
-				rows[i].name = malloc(strlen(ep->d_name) + 2);
+				rows[i].name = (char *) malloc(strlen(ep->d_name) +2);
 				strcpy(rows[i].name, ep->d_name);
 				if (!rows[i].islink)
-					strcat(rows[i].name, "/");
+					ADDSLASH(rows[i].name);
 				rows[i].mode = statbuf.st_mode;
 				i++;
 			}
 		} else if (flags & SHOW_FILES) {
-			rows[i].name = malloc(strlen(ep->d_name) + 1);
+			rows[i].name = (char *) malloc(strlen(ep->d_name) +1);
 			strcpy(rows[i].name, ep->d_name);
 			rows[i].size = statbuf.st_size;
 			rows[i].mode = statbuf.st_mode;
@@ -405,6 +290,7 @@ int ls(Row **rowsp, uint8_t flags)
 	qsort(rows, n, sizeof(*rows), rowcmp);
 	closedir(dp);
 	*rowsp = rows;
+	
 	return n;
 }
 
@@ -413,44 +299,50 @@ void free_rows(Row **rowsp, int nfiles)
 	int i;
 
 	for (i = 0; i < nfiles; i++)
-		free((*rowsp)[i].name);
-	free(*rowsp);
-	*rowsp = NULL;
+		FREE((*rowsp)[i].name);
+	FREE(*rowsp);
 }
 
 /* Change working directory to the path in CWD. */
-void cd(int reset)
+void cd(bool reset)
 {
 	int i, j;
 
 	message(CYAN, "Loading \"%s\"...", CWD);
 	refresh();
 	if (chdir(CWD) == -1) {
-		getcwd(CWD, PATH_MAX - 1);
-		if (CWD[strlen(CWD) - 1] != '/')
-			strcat(CWD, "/");
-		goto done;
+		getcwd(CWD, PATH_MAX -1);
+		ADDSLASH(CWD);
+
+		clear_message();
+		update_view();
+
+		return;
 	}
+
 	if (reset)
 		ESEL = SCROLL = 0;
+
 	if (rover.nfiles)
 		free_rows(&rover.rows, rover.nfiles);
+
 	rover.nfiles = ls(&rover.rows, FLAGS);
 	if (!strcmp(CWD, rover.marks.dirpath)) {
 		for (i = 0; i < rover.nfiles; i++) {
 			for (j = 0; j < rover.marks.bulk; j++)
-				if (
-					rover.marks.entries[j] &&
-					!strcmp(rover.marks.entries[j], ENAME(i)))
+				if ( rover.marks.entries[j] && !strcmp(rover.marks.entries[j], ENAME(i)))
 					break;
+
 			MARKED(i) = j < rover.marks.bulk;
 		}
 	} else
 		for (i = 0; i < rover.nfiles; i++)
 			MARKED(i) = 0;
-done:
+
 	clear_message();
 	update_view();
+
+	return;
 }
 
 /* Select a target entry, if it is present. */
@@ -471,11 +363,11 @@ void reload()
 
 	if (rover.nfiles) {
 		strcpy(input, ENAME(ESEL));
-		cd(0);
+		cd(false);
 		try_to_sel(input);
 		update_view();
 	} else
-		cd(1);
+		cd(true);
 }
 
 off_t count_dir(const char *path)
@@ -664,7 +556,7 @@ int cpyfile(const char *srcpath)
 		if (ret < 0)
 			return ret;
 		buffer[ret] = '\0';
-		ret       = symlink(buffer, dstpath);
+		ret         = symlink(buffer, dstpath);
 	} else {
 		ret = src = open(srcpath, O_RDONLY);
 		if (ret < 0)
@@ -724,10 +616,10 @@ void start_line_edit(const char *init_input)
 
 	curs_set(TRUE);
 	strncpy(input, init_input, PATH_MAX);
-	rover.edit.left           = mbstowcs(rover.edit.buffer, init_input, PATH_MAX);
-	rover.edit.right          = PATH_MAX - 1;
+	rover.edit.left             = mbstowcs(rover.edit.buffer, init_input, PATH_MAX);
+	rover.edit.right            = PATH_MAX - 1;
 	rover.edit.buffer[PATH_MAX] = L'\0';
-	rover.edit_scroll         = 0;
+	rover.edit_scroll           = 0;
 }
 
 /* Read input and change editing state accordingly. */
@@ -788,4 +680,3 @@ EditStat get_line_edit()
 	         PATH_MAX - length);
 	return CONTINUE;
 }
-
